New restore role
There is a new restore role in the roles folder:

    roles/iparestore

This role allows to restore an IPA server locally and from the controller
and also to copy a backup from the controller to the server.

Here is the documentation for the role:

    roles/ipabackup/README.md

New example playbooks have been added:

    playbooks/restore-server.yml
    playbooks/restore-server-from-controller.yml
    playbooks/copy-backup-to-server.yml

The vars are copied form the ipaserver role to be able to install the
needed packages. The vars should be moved into a common role to be consumed
by ipaserver, ipareplica and iparestore roles later on.