Allow user override of distro specific variables.
This patch adds variable *_user_vars to backup, server, replica and
client roles, allowing a user to provide a file to be used instead of
the distro specific ones.

This change allows usage of ansible-freeipa in non-supported distros
without requiring the addition of a distro specific file in the role
`vars` directory.