ansible-freeipa-test.sh: Testing modules locally with containers.
This patch provides a way to run ansible-freeipa test playbooks with
containers and software tools installed in a virtual environment. The
script is designed to run on the developer's workstation, using the
same test images an used in the project CI environment.

A new utility can be found at

    util/ansible-freeipa-test.sh

Documentation for using it for testing can be found at:

    tests/README-containers.md