ansible_freeipa_module: Allow api_connect configuration.
The function api_connect() is used to bootstrap the IPA API, and
configure the connection to the backend in this process. Currently,
only the `context` in which the commands will be executed, but other
attributes can also be set, like enabling/disabling LDAP cache layer.

By modifying ansible_freeipa_module.api_connect() to accept a set of
keyword arguments, it is possible to configure the bootstraping
process with the desired behavior.

The keyword arguments then have their type and values validated. Also,
if required, value mapping is applied, and, if not present, sensible
default values are provided.

If the argument type is invalid, or if an unexpected argument is given,
a TypeError is raised. If the value provided for the argument is not
in the list of accepted values a ValueError is raised.

Currently, the valid arguments are:
  - `context`: The context to run the module, accepts 'server',
    'client',or 'cli', the default is 'server', and 'client' is mapped
    to 'cli'.
  - `ldap_cache`: A bool value turning on/off the use LDAP caching
    layer for executing commands. Default is `True`.