Allow execution of plugins in clients hosts.
By modifying ansible_freeipa_module.api_connect() to accept a string
defining a context to execute the plugins, an user can configure an
IPA server through a client host, thus not requiring ssh access to
the server.

The context string should contain only "server" for execution in a
server host, or "client" for execution in a client host. If not
provided, the default value is "server". Internally, IPA uses "cli"
as the context name for client plugins, and it is also accepted.

A ValueError exception is raised if an invalid context value is
provided.