ipadnsforwardzone: Allow execution of plugins in clients hosts.
Add a new configuration variable, `ipa_context` to module
ipadnsforwardzone, to allow an user to select the environment
context the module will execute, server or client.

A new test playbook is added:

    tests/dnsforwardzone/test_dnsforwardzone_client_context.yml

As the test requires both server and client hosts to be available,
the tasks are wrapped in a block that only executes if
`group['ipaclients']` is defined through an inventory file.