topology: Allow execution of plugins in clients hosts.
Add a new configuration variable, `ipa_context` to modules
ipatopologysegment and ipatopologysuffix, to allow an user to select
the environment context the module will execute, server or client.

The module ipatopologysegment has been modified to use the new
higher level connection API, so that its code is more consistent
with other modules and future changes don't have to be implemented
only for this module, that was the only one using execute_api_command.

As there are currently no test playbooks for topology modules, the
tests were performed using modified example playbooks.