module_utils: Add option to configure ldap cache usage.
FreeIPA 4.9.4 has introduced a short term, per-connection LDAP cache
layer aiming to reduce load on the LDAP server by caching similar
queries executed in the same connection, and, although it works well
for most ansible-freeipa modules, the use of the cache layer may
trigger code paths, in FreeIPA, that don't work well with the way
queries are cached.

This patch allows a module to provide extra parameters to
api_connect(), allowing, for example, to configure if the ldap cache
is to be used or not.