FreeIPABaseModule: add support for 'delete_continue'.
Many FreeIPA objects provide support for 'continuous mode' when
deleting objects, and this behavior is very similar in many objects.

This change allows modules implemented using FreeIPABaseModule to
leverage the same interface for this attribute simply passing
`support_delete_continue=True` to the FreeIPABaseModule constructor.

It is still the responsibility of the derived class properly check
for correct usage of the attribute. A new method
`check_invalid_params` is provided to help with this check.