ansible_freeipa_module: Helper to filter services.
Services may be defined in playbooks without the realm and this may
cause comparision to fail, since IPA stores services with '@REALM'.

This patch adds a function to ansible_freeipa_module that creates a
list of services that matches a given predicate. The predicate is
invoked with a normalized service name (with '@REALM') and a list
of normalized service names, so the creation of add/del list for
services can be created without the need to verify the existence of
the realm part.