sudorule: case insensitive comparision of members.
When having a non-normalized list of hosts, hostgroups, users, groups,
runasuser or runasgroup, including names with upper and lowercase, the
comparision with existing members fail, as members are always added in
lowercase.

This patch ensures required data is stored in lowercase.

A new test file  was added to verify the issue:

   tests/sudorule/test_sudorule_member_case_insensitive.yml