ipavault: Fix idempotent issues with members.
When members are used with the `ipavault` module, the task returns
'changed: true', even if the same task is executed twice.

This patch fixes this issue, by using gen_add_lists and
gen_del_lists when using 'action: member', and by only trying to
create the member lists if the member was actually provided.

The service atttributes (service, services, and ownerservice),
may be modified to ensure that the realm is present.

The user 'admin' required special handling, and is only removed if
explicitly set to be removed. This was necessary and the result of
running the same task resulted in different vault states.

Also, the command generation has been refactored, and is now coded
only once, after membership has been defined for all attributes.